removed authentication
It looks like a totally different approach is needed: it cannot really be
done as a separate service rather as an add-on for the main Calibre installation.
The root cause is that it is simply impossible to maintain two separate book
database and syncing would be a mess.

This is the first step towards using this as an addon on top of the Calibre
content server.